SMT32 work
Some SMT32 work and decrease RAM usage on ARM

+++ speeduino/cancomms.ino
@@ -57,14 +57,14 @@ void canCommand()
        break;
 
     case 'S': // send code version
-       for (unsigned int sig = 0; sig < sizeof(displaySignature) - 1; sig++){
+       for (uint16_t sig = 0; sig < sizeof(displaySignature) - 1; sig++){
            Serial3.write(displaySignature[sig]);
        }
        //Serial3.print("speeduino 201609-dev");
        break;
 
     case 'Q': // send code version
-       for (unsigned int revn = 0; revn < sizeof( TSfirmwareVersion) - 1; revn++){
+       for (uint16_t revn = 0; revn < sizeof( TSfirmwareVersion) - 1; revn++){
            Serial3.write( TSfirmwareVersion[revn]);
        }
        //Serial3.print("speeduino 201609-dev");
