+++ speeduino/comms.h
@@ -33,8 +33,8 @@ const char pageTitles[] PROGMEM //This is being stored in the avr flash instead 
   };
 
 void command();//This is the heart of the Command Line Interpeter.  All that needed to be done was to make it human readable.
-void sendValues(int packetlength, byte portnum);
-void receiveValue(int offset, byte newValue);
+void sendValues(int16_t packetlength, byte portnum);
+void receiveValue(int16_t offset, byte newValue);
 void saveConfig();
 void sendPage(bool useChar);
 void receiveCalibration(byte tableID);
