+++ speeduino/table.h
@@ -19,11 +19,11 @@ struct table2D {
   byte *values;
   byte *axisX;
   
-  int *values16;
-  int *axisX16;
+  int16_t *values16;
+  int16_t *axisX16;
   
   //Store the last X and Y coordinates in the table. This is used to make the next check faster
-  int lastXMax, lastXMin;
+  int16_t lastXMax, lastXMin;
 };
 
 void table2D_setSize(struct table2D targetTable, byte newSize);
@@ -36,8 +36,8 @@ struct table3D {
   byte ySize;
   
   byte **values;
-  int *axisX;
-  int *axisY;
+  int16_t *axisX;
+  int16_t *axisY;
   
   //Store the last X and Y coordinates in the table. This is used to make the next check faster
   byte lastXMax, lastXMin;
@@ -59,7 +59,7 @@ Eg: 2x2 table
 (1,0) = 1
 
 */
-int get3DTableValue(struct table3D *fromTable, int, int);
-int table2D_getValue(struct table2D *fromTable, int);
+int16_t get3DTableValue(struct table3D *fromTable, int16_t, int16_t);
+int16_t table2D_getValue(struct table2D *fromTable, int16_t);
 
 #endif // TABLE_H
